my ordered page completed

+++ src/Components/MyAddedCard.jsx
@@ -1,24 +1,45 @@
 import { Link } from "react-router-dom";
 import PropTypes from 'prop-types';
+import Swal from "sweetalert2";
 
-const MyAddedCard = ({card}) => {
-    const {food_name , price , food_origin , purchase_count , quantity , _id ,photo_url
+const MyAddedCard = ({card , reload, setReload}) => {
+    const {food_name , price , food_origin, purchase_count , quantity , _id ,photo_url
     }= card || {}
 
 
     const handleDelete=_id=>{
         console.log('delete', _id);
-        fetch(`http://localhost:3000/delete/${_id}`, {
+
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            fetch(`http://localhost:3000/delete/${_id}`, {
           method: "DELETE",
         })
           .then((res) => res.json())
           .then((data) => {
             console.log(data);
             if (data.deletedCount) {
-              // location.reload()
-              alert("deleted")
+              setReload(!reload)
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success"
+              });
             }
           });
+          }
+        });
+
+
+        
     }
     return (
         <div>
@@ -72,6 +93,8 @@ const MyAddedCard = ({card}) => {
     );
 };
 MyAddedCard.propTypes = {
-    card: PropTypes.object
+    card: PropTypes.object,
+    reload:PropTypes.bool,
+    setReload: PropTypes.func
   };
 export default MyAddedCard;
